+++ pcs_test/tier0/common/reports/test_messages.py
@@ -6351,11 +6351,26 @@ class NoStonithMeansWouldBeLeft(NameBuildTest):
     def test_success(self):
         self.assert_message_from_report(
             (
-                "Requested action lefts the cluster with no enabled means "
+                "Requested action leaves the cluster with no enabled means "
                 "to fence nodes, resulting in the cluster not being able to "
                 "recover from certain failure conditions"
             ),
             reports.NoStonithMeansWouldBeLeft(),
+        )
+
+
+class NoStonithMeansWouldBeLeftDueToProperties(NameBuildTest):
+    def test_success(self):
+        self.assert_message_from_report(
+            (
+                "Setting property stonith-enabled to false or fencing-enabled"
+                " to 0 leaves the cluster with no enabled means to fence nodes,"
+                " resulting in the cluster not being able to recover from"
+                " certain failure conditions"
+            ),
+            reports.NoStonithMeansWouldBeLeftDueToProperties(
+                {"stonith-enabled": "false", "fencing-enabled": "0"}
+            ),
         )